Move workspaces-preview directory to batch-spec/edit
